+++ checkpoint/CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Allow saving the aggregated file asynchronously.
 
+### Fixed
+- Support creating sharded array when ArrayRestoreArgs is passed and the value
+was originally aggregated.
+
 ## [0.2.5] - 2022-06-09
 
 ### Added